stt/requirements.txt
# Web
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ML core (КРИТИЧНО)
numpy==1.26.4
torch==2.1.2
torchaudio==2.1.2

# WhisperX stack (ПРОВЕРЕНО)
whisperx==3.1.1
pyannote.audio==3.1.1

# Utils
soundfile
python-multipart
ffmpeg-python
